refactor: changed the way attendance handles functions for better understanding of logic.
