--- src/pages/attendanceview.tsx
@@ -1,10 +1,0 @@
-import CsvViewer from "../../views/AttendanceProcessor";
-import { Box } from "@chakra-ui/react/box";
-
-const AttendanceViewPage = () => {
-  return (
-      <CsvViewer/>
-  );
-};
-
-export default AttendanceViewPage;
